% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dislib-package.R
\docType{package}
\name{dislib-package}
\alias{dislib}
\alias{dislib-package}
\title{dislib: Give insight about the diseases}
\description{
This package contains a data set and function to provide information on mortality rate for each diseases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etc5523-2022/rpkg-KrisanatA}
  \item Report bugs at \url{https://github.com/etc5523-2022/rpkg-KrisanatA/issues}
}

}
\author{
\strong{Maintainer}: Krisanat Anukarnsakulchularp \email{kanu0003@student.monash.edu}

}
\keyword{internal}
