% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Launch a shiny app}
\usage{
run_app()
}
\value{
Shiny app.
}
\description{
This function is used to called the shiny app that are based on the provided dataset \code{mortality}.
}
