% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{year_list}
\alias{year_list}
\title{Year list}
\usage{
year_list(country = NULL)
}
\arguments{
\item{country}{A character of the country name you want to learn about the disease's rank. See a full list of countries by \code{country_list()}.}
}
\value{
A vector of available year value for a particular country.
}
\description{
Get a list of the available year for the country specified.
}
\examples{
year_list(country = "Australia")

}
