% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis-summary.R
\name{dis_summary}
\alias{dis_summary}
\title{Disease summary}
\usage{
dis_summary(country = NULL, year = NULL)
}
\arguments{
\item{country}{A character of country name you want to learn about the diseases rank. See a full list of country by \code{country_list()}.}

\item{year}{A numeric for a specific year of interest. See a full list of year by \code{year_list()}.}
}
\value{
A data frame on the top 10 diseases mortality rate from highest to lowest.
}
\description{
Rank the highest mortality rate by disease in a particular country and year.
For a list of available country and year refer back to \code{\link[=country_list]{country_list()}}, \code{\link[=year_list]{year_list()}}.
}
\examples{
dis_summary(country = "Australia", year = 2019)

}
