% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{year_list}
\alias{year_list}
\title{Year list}
\usage{
year_list(country = NULL)
}
\arguments{
\item{country}{A character of country name you want to learn about the diseases rank. See a full list of country by \code{country_list()}.}
}
\value{
A vector of available year data for particular country.
}
\description{
Get a list of available year for the country specified.
}
\examples{
year_list(country = "Australia")

}
