% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{year_list}
\alias{year_list}
\title{Get a list of available year for the country specified}
\usage{
year_list(country = NULL)
}
\arguments{
\item{country}{A character of country name}
}
\description{
Get a list of available year for the country specified
}
